refactor with es6 modules
